add view Smart class
